requirements.txt
faker
reportlab
datasets 
transformers 
pytorch-lightning
pymupdf 
pytesseract 
pillow 
pandas
torchvision 
sentencepiece 
timm
pdf2image
reportlab   
opencv-python 
pytesseract 
numpy  
scikit-learn 
spacy 
tqdm